refactor webpack config. Now routing works as expected
