+++ package.json
@@ -5,11 +5,11 @@
   "scripts": {
     "clean": "rimraf ./src/*.js ./src/*.js.map && npm cache clean",
     "clean-install": "npm run clean && npm install",
-    "start": "webpack-dev-server  --progress --colors --port 3100 --inline --config webpack.config.js",
+    "start": "webpack-dev-server  --progress --display-error-details  --colors --port 3100 --inline --config webpack.config.js --content-base src",
     "start.prod": "webpack --config config/webpack.prod.js  --progress --profile --colors --display-error-details --display-cached --bail",
     "clean-start": "npm run clean && npm start.dev",
     "watch": "webpack --watch --progress --profile",
-    "build": "rimraf dist  && webpack -d --progress --profile",
+    "build": "rimraf dist  && webpack -d --progress --profile --config webpack.config.js",
     "deploy": "webpack -p",
     "docker-build": "docker build -t ng2-quickstart .",
     "docker": "npm run docker-build && docker run -it --rm -p 3000:3000 -p 3001:3001 ng2-quickstart",
@@ -27,57 +27,31 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@angular/common": "^2.0.0-rc.5",
-    "@angular/compiler": "^2.0.0-rc.5",
-    "@angular/core": "^2.0.0-rc.5",
-    "@angular/forms": "^0.3.0",
-    "@angular/http": "^2.0.0-rc.5",
-    "@angular/platform-browser": "^2.0.0-rc.5",
-    "@angular/platform-browser-dynamic": "^2.0.0-rc.5",
-    "@angular/platform-server": "^2.0.0-rc.5",
-    "@angular/router": "^3.0.0-rc.1",
-    "@angular2-material/button": "^2.0.0-alpha.7-4",
-    "@angular2-material/button-toggle": "^2.0.0-alpha.7-4",
-    "@angular2-material/card": "^2.0.0-alpha.7-4",
-    "@angular2-material/checkbox": "^2.0.0-alpha.7-4",
-    "@angular2-material/core": "^2.0.0-alpha.7-4",
-    "@angular2-material/grid-list": "^2.0.0-alpha.7-4",
-    "@angular2-material/icon": "^2.0.0-alpha.7-4",
-    "@angular2-material/input": "^2.0.0-alpha.7-4",
-    "@angular2-material/list": "^2.0.0-alpha.7-4",
-    "@angular2-material/menu": "^2.0.0-alpha.7-4",
-    "@angular2-material/progress-bar": "^2.0.0-alpha.7-4",
-    "@angular2-material/progress-circle": "^2.0.0-alpha.7-4",
-    "@angular2-material/radio": "^2.0.0-alpha.7-4",
-    "@angular2-material/sidenav": "^2.0.0-alpha.7-4",
-    "@angular2-material/slide-toggle": "^2.0.0-alpha.7-4",
-    "@angular2-material/slider": "^2.0.0-alpha.7-4",
-    "@angular2-material/tabs": "^2.0.0-alpha.7-4",
-    "@angular2-material/toolbar": "^2.0.0-alpha.7-4",
-    "@angular2-material/tooltip": "^2.0.0-alpha.7-4",
-    "@angularclass/conventions-loader": "^1.0.12",
-    "@angularclass/request-idle-callback": "^1.0.7",
-    "@angularclass/webpack-toolkit": "^1.3.3",
+    "@angular/common": "2.0.0-rc.5",
+    "@angular/compiler": "2.0.0-rc.5",
+    "@angular/core": "2.0.0-rc.5",
+    "@angular/forms": "0.3.0",
+    "@angular/http": "2.0.0-rc.5",
+    "@angular/platform-browser": "2.0.0-rc.5",
+    "@angular/platform-browser-dynamic": "2.0.0-rc.5",
+    "@angular/platform-server": "2.0.0-rc.5",
+    "@angular/router": "3.0.0-rc.1",
     "angular2-jwt": "0.1.18",
-    "angular2-mdl": "^1.4.6",
-    "core-js": "^2.4.0",
-    "hammerjs": "^2.0.8",
-    "ie-shim": "^0.1.0",
-    "moment": "^2.13.0",
-    "reflect-metadata": "^0.1.4",
-    "rxjs": "^5.0.0-beta.11",
-    "zone.js": "^0.6.12"
+    "angular2-mdl": "1.4.6",
+    "core-js": "2.4.0",
+    "ie-shim": "0.1.0",
+    "moment": "2.13.0",
+    "rxjs": "5.0.0-beta.11",
+    "zone.js": "0.6.12"
   },
   "devDependencies": {
+    "angular2-template-loader": "^0.5.0",
+    "awesome-typescript-loader": "^1.1.1",
     "browser-sync": "^2.13.0",
     "browser-sync-webpack-plugin": "^1.0.3",
-    "canonical-path": "0.0.2",
-    "copy-webpack-plugin": "^3.0.1",
     "css-loader": "^0.23.1",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
-    "html-loader": "^0.4.3",
-    "html-webpack-plugin": "^2.22.0",
     "http-server": "^0.9.0",
     "jasmine-core": "~2.4.1",
     "json-server": "^0.8.17",
@@ -86,21 +60,17 @@
     "karma-cli": "^0.1.2",
     "karma-htmlfile-reporter": "^0.2.2",
     "karma-jasmine": "^0.3.8",
-    "lodash": "^4.11.1",
     "protractor": "^3.3.0",
-    "raw": "^0.1.4",
     "raw-loader": "^0.5.1",
     "rimraf": "^2.5.2",
-    "style-loader": "^0.13.1",
-    "ts-loader": "^0.8.2",
+    "to-string-loader": "^1.1.4",
     "tslint": "^3.7.4",
     "typedoc": "^0.4.4",
     "typescript": "^1.8.10",
     "typings": "^1.0.4",
-    "url-loader": "^0.5.7",
     "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-dotenv-plugin": "^1.3.2"
+    "webpack-dotenv-plugin": "^1.3.2",
+    "webpack-merge": "^0.14.1"
   },
   "repository": {}
 }
