+++ src/app/shared/common/Entry.ts
