+++ src/app/shared/common/entry.interface.ts
