+++ src/app/shared/index.ts
