+++ src/app/yearly/index.ts
