+++ src/app/yearly/yearly.component.ts
