+++ src/index.html
@@ -57,7 +57,7 @@
 
         html, body {
             font-family: 'Roboto', 'Helvetica', sans-serif;
-            margin: 0;
+            margin: 0;^
             padding: 0;
             width: 100%;
             height: 100%;
@@ -69,5 +69,10 @@
 <body class="mdl-demo mdl-color--grey-100 mdl-color-text--grey-700 mdl-base">
 <mdl-textfield type="text" [(ngModel)]="theText" icon="search"></mdl-textfield>
 <sleep-diary>Loading...</sleep-diary>
+
+<script src="/polyfills.bundle.js"></script>
+<script src="/vendor.bundle.js"></script>
+<script src="/main.bundle.js"></script>
+
 </body>
 </html>
