+++ tsconfig.json
@@ -1,21 +1,21 @@
 {
   "compileOnSave": true,
   "compilerOptions": {
-    "declaration": false,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "mapRoot": "/",
     "module": "commonjs",
-    "moduleResolution": "node",
-    "noEmitOnError": true,
-    "outDir": "../dist/",
+    "target": "es5",
+    "outDir": "dist",
     "rootDir": ".",
     "sourceMap": true,
-    "target": "es5",
-    "inlineSources": true
+    "emitDecoratorMetadata": true,
+    "experimentalDecorators": true,
+    "moduleResolution": "node"
   },
-  "files": [
-    "./app/main.ts",
-    "./app/typings.d.ts"
-  ]
+  "exclude": [
+    "node_modules"
+  ],
+  "awesomeTypescriptLoaderOptions": {
+    "useWebpackText": true
+  },
+  "compileOnSave": false,
+  "buildOnSave": false
 }
