+++ typings.json
@@ -4,7 +4,7 @@
     "core-js": "registry:dt/core-js#0.0.0+20160725163759",
     "hammerjs": "registry:dt/hammerjs#2.0.8+20160724063812",
     "jasmine": "registry:dt/jasmine#2.2.0+20160621224255",
-    "node": "registry:dt/node#6.0.0+20160807145350",
+    "node": "registry:dt/node#6.0.0+20160831021119",
     "selenium-webdriver": "registry:dt/selenium-webdriver#2.44.0+20160317120654"
   }
 }
